+++ src/components/separators/iframe-pages/small-device.jsx
